working refactor of init_section
