+++ stellar-engine/src/stellar_engine/utils.py
@@ -1,0 +1,3 @@
+
+def get_section_middle_span(start_support: int, end_support: int):
+    return (start_support + end_support) // 2
